docs: bring showcases back
