+++ docs/showcase/markdown.mdx
@@ -3,7 +3,7 @@ hide_title: true
 title: Markdown Editor
 ---
 
-import { Example } from '@components';
+import Basic from '../../website/extension-examples/react-editors/markdown-editor';
 
 # Markdown Editor
 
@@ -12,6 +12,8 @@ The `<MarkdownEditor>` provides a state of the art editor to edit markdown.
 ```ts
 import { MarkdownEditor } from '@remirror/react-editors/markdown';
 ```
+
+<Basic />
 
 Find many more examples on [Storybook](https://remirror.vercel.app/?path=/story/editors-markdown--basic).
