+++ docs/showcase/richtext.mdx
@@ -3,7 +3,7 @@ hide_title: true
 title: Wysiwyg Editor
 ---
 
-import { Example } from '@components';
+import Basic from '../../website/extension-examples/react-editors/wysiwyg-editor';
 
 # Wysiwyg Editor
 
@@ -12,6 +12,8 @@ The `<WysiwygEditor>` provides a state of the art richtext editor.
 ```ts
 import { WysiwygEditor } from '@remirror/react-editors/wysiwyg';
 ```
+
+<Basic />
 
 Find many more examples on [Storybook](https://remirror.vercel.app/?path=/story/editors-wysiwyg--basic).
