+++ docs/showcase/social.mdx
@@ -3,7 +3,7 @@ hide_title: true
 title: Social Editor
 ---
 
-import { Example } from '@components';
+import Basic from '../../website/extension-examples/react-editors/social-editor';
 
 # Social Editor
 
@@ -12,6 +12,8 @@ The `<SocialEditor>` provides a state of the art editor with mention and tag sup
 ```ts
 import { SocialEditor } from '@remirror/react-editors/social';
 ```
+
+<Basic />
 
 Find many more examples on [Storybook](https://remirror.vercel.app/?path=/story/editors-social--basic).
